requirements.txt
flask
werkzeug
scikit-learn
google-generativeai
faiss-cpu
numpy
regex
